[feat] CryptoManager 생성 (#96)
